Use int instead of size_t in BitReader and ByteOrderDecoder
Migrate lengths, offsets, and bit counts to int. Change ParseSimpleArchive dataLen to int and drop (size_t) casts at call sites. Fix BitReader::Eat comparing against the nBits parameter instead of total stream bits.

prompt: change BitReader and ByteOrderDecoder to use int instead of size_t; propagate changes to callers to simplify int <=> size_t casting

+++ src/CrashHandler.cpp
@@ -268,7 +268,7 @@ void UploadCrashReport(Str d) {
 static bool ExtractSymbols(Str archiveData, Str dstDir, Arena* allocator) {
     logf("ExtractSymbols: dir '%s', size: %d\n", dstDir, archiveData.len);
     lzma::SimpleArchive archive;
-    bool ok = ParseSimpleArchive((const u8*)archiveData.s, (size_t)archiveData.len, &archive);
+    bool ok = ParseSimpleArchive((const u8*)archiveData.s, archiveData.len, &archive);
     if (!ok) {
         logf("ExtractSymbols: ParseSimpleArchive failed\n");
         return false;
